+++ CHANGELOG.md
@@ -44,7 +44,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - LASRV001 - 25/03/2023 - Quit CPUException in not important information of CPU.
 - LASRV008 - 28/03/2023 - Change the manager exceptions to give a standard description.
 - LASRV008 - 29/03/2023 - Refactor the exception message of the ManagerException.
-- LASRV008 - 29/03/2023 - Add components enum in CPUInfo.
+- LASRV001 - 29/03/2023 - Add components enum in CPUInfo.
+- LASRV002 - 29/03/2023 - Add components enum in RAMInfo.
 
 ### Fixed
 - LASRV006 - 26/03/2023 - Fix add user bug.